patch security problem

+++ src/main/java/com/adamfgcross/springherokudemo/controller/DropFormController.java
@@ -15,6 +15,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.util.Collections;
 import java.util.List;
@@ -22,6 +23,10 @@ import java.util.Optional;
 
 @Controller
 public class DropFormController {
+
+    @ResponseStatus(value = HttpStatus.NOT_FOUND)
+    public static class ResourceNotFoundException extends RuntimeException {
+    }
 
     @Autowired
     private DropService dropService;
@@ -37,7 +42,7 @@ public class DropFormController {
     }
 
     @GetMapping("/drop/{id}")
-    public String getDrop(@PathVariable Long id, Model model) {
+    public String getDrop(@PathVariable Long id, Model model, HttpServletResponse httpServletResponse) {
         Optional<Drop> dropOptional = dropService.findById(id);
         Drop drop;
         if (dropOptional.isPresent()) {
@@ -45,7 +50,7 @@ public class DropFormController {
             model.addAttribute("update",true);
 
         } else {
-            throw new RuntimeException("could not find drop");
+            throw new ResourceNotFoundException();
         }
         model.addAttribute("drop", drop);
         return "drop-form";
@@ -58,24 +63,67 @@ public class DropFormController {
             System.out.println(bindingResult.getAllErrors());
             return "drop-form";
         }
+        // check if post is an update
+        Boolean isUpdate = false;
+        if (drop.getId() != null) {
+            isUpdate = true;
+        }
+
 
         String dropText = drop.getText();
         Optional<User> userOptional = getUser(authentication);
-        System.out.println("received dropText" + dropText);
+        System.out.println("received dropText: " + dropText);
         if (!userOptional.isPresent()) {
+            // this should never happen
             redirAttrs.addFlashAttribute("errorText", "User not found in database");
         } else {
             User user = userOptional.get();
-            drop.setUser(user);
-            dropService.saveDrop(drop);
-            //dropService.saveDrop(dropText, user);
-            redirAttrs.addFlashAttribute("messageText", "Drop saved successfully");
+            if (isUpdate) {
+                Long dropId = drop.getId();
+                Long userId = user.getId();
+                Optional<Drop> existingDrop = dropService.findById(dropId);
+                if (!existingDrop.isPresent()) {
+                    // this should never happen
+                    redirAttrs.addFlashAttribute("errorText", "Could not find existing drop data.");
+                    return "redirect:/drop";
+                } else {
+                    if (existingDrop.get().getUser().getId().equals(userId)) {
+                        // user is authorized
+                        drop.setUser(user);
+                        dropService.saveDrop(drop);
+                        redirAttrs.addFlashAttribute("messageText", "Drop saved successfully");
+                    } else {
+                        // edit/update was unauthorized
+                        redirAttrs.addFlashAttribute("errorText", "An error occurred. Please try again.");
+                    }
+                }
+            } else {
+                drop.setUser(user);
+                dropService.saveDrop(drop);
+                redirAttrs.addFlashAttribute("messageText", "Drop saved successfully");
+            }
         }
         return "redirect:/drop";
     }
 
     @DeleteMapping("/drop/{id}")
-    public ResponseEntity<Long> deleteDrop(@PathVariable Long id) {
+    public ResponseEntity<Long> deleteDrop(@PathVariable Long id, Authentication authentication) {
+        Optional<User> userOptional =  getUser(authentication);
+        Optional<Drop> dropOptional = dropService.findById(id);
+        if (!userOptional.isPresent()) {
+            throw new RuntimeException("user not found");
+        }
+        if (!dropOptional.isPresent()) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        Drop drop = dropOptional.get();
+        User user = userOptional.get();
+        Long userId = user.getId();
+        // check if the drop belongs to the user
+        if (drop.getUser().getId() != userId) {
+            // if not, return forbidden
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         if (dropService.removeDrop(id)) {
             return ResponseEntity.ok(id);
         } else {
